requirements.txt
pdfplumber
PyYAML
python-docx
openpyxl
virtualenv